fex: Semantic and Variable class separation
